[drivers] Move the shared FFM plumbing into drivers-common and grant it native access
