+++ tornado-drivers/cuda/src/main/java/uk/ac/manchester/tornado/drivers/cuda/CUDAEvent.java
@@ -43,7 +43,7 @@ import uk.ac.manchester.tornado.drivers.cuda.enums.CUDAProfilingInfo;
 import uk.ac.manchester.tornado.drivers.cuda.exceptions.CUDAException;
 import uk.ac.manchester.tornado.drivers.cuda.ffm.CUDADriverAPI;
 import uk.ac.manchester.tornado.drivers.cuda.ffm.CUDAHandles;
-import uk.ac.manchester.tornado.drivers.cuda.ffm.FFMSupport;
+import uk.ac.manchester.tornado.drivers.common.ffm.FFMSupport;
 import uk.ac.manchester.tornado.runtime.common.RuntimeUtilities;
 import uk.ac.manchester.tornado.runtime.common.TornadoLogger;
