+++ tornado-drivers/cuda/src/main/java/uk/ac/manchester/tornado/drivers/cuda/CUDAKernel.java
@@ -35,7 +35,7 @@ import uk.ac.manchester.tornado.drivers.cuda.enums.CUDAKernelInfo;
 import uk.ac.manchester.tornado.drivers.cuda.exceptions.CUDAException;
 import uk.ac.manchester.tornado.drivers.cuda.ffm.CUDADriverAPI;
 import uk.ac.manchester.tornado.drivers.cuda.ffm.CUDAHandles;
-import uk.ac.manchester.tornado.drivers.cuda.ffm.FFMSupport;
+import uk.ac.manchester.tornado.drivers.common.ffm.FFMSupport;
 import uk.ac.manchester.tornado.runtime.common.TornadoLogger;
 
 public class CUDAKernel {
